refactored case summary tests
